requirements.txt
python-dotenv~=1.0.1
pyyaml~=6.0.2
urllib3~=1.26.15
mysql-connector-python~=9.1.0